Move client assets next to the files that use them
This is closer to idiomatic webpack style. Moves assets used by more
than one package into /client/shared-res
