+++ src/client/dashboard/res/barberpole-error.png
