+++ src/client/dashboard/res/character-menu-arrow.png
