+++ src/client/roster/res/search-roster.png
